+++ CHANGELOG.md
@@ -6,9 +6,17 @@ Please, document here only changes visible to the client app.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2026-03-03
+
+### [47 Better Error Handling](https://github.com/torqlab/torq/issues/47)
+
+### Added
+
+- Better error handling to the Strava API Client
+
 ## [1.1.0] - 2026-03-03
 
-### [47 Export TypeScript Types for External Use](https://github.com/torqlab/strava-api/issues/47)
+### [47 Export TypeScript Types for External Use](https://github.com/torqlab/torq/issues/47)
 
 ### Added